Fixed a few compiler warnings

+++ skFixedString.h
@@ -136,7 +136,7 @@ public:
 
     void append(const char* str)
     {
-        int len = strlen(str);
+        int len = (int)strlen(str);
         int a   = 0;
         while (a < len)
             push_back(str[a++]);
